+++ dio_test/lib/src/test/parameter_tests.dart
@@ -1,0 +1,46 @@
+import 'package:dio/dio.dart';
+import 'package:test/test.dart';
+
+import '../utils.dart';
+
+void parameterTests(
+  Dio Function() create,
+) {
+  late Dio dio;
+
+  setUpAll(() {
+    dio = create();
+  });
+
+  group('parameters', () {
+    group('generic parameters', () {
+      test('default (Map)', () async {
+        final response = await dio.get('/get');
+        expect(response.data, isA<Map>());
+        expect(response.data, isNotEmpty);
+      });
+
+      test('Map', () async {
+        final response = await dio.get<Map>('/get');
+        expect(response.data, isA<Map>());
+        expect(response.data, isNotEmpty);
+      });
+
+      test('String', () async {
+        final response = await dio.get<String>('/get');
+        expect(response.data, isA<String>());
+        expect(response.data, isNotEmpty);
+      });
+
+      test('List', () async {
+        final response = await dio.post<List>(
+          '/payload',
+          data: '[1,2,3]',
+        );
+        expect(response.data, isA<List>());
+        expect(response.data, isNotEmpty);
+        expect(response.data![0], 1);
+      });
+    });
+  });
+}
