+++ dio_test/lib/tests.dart
@@ -1,0 +1,5 @@
+export 'src/test/headers_tests.dart';
+export 'src/test/http_method_tests.dart';
+export 'src/test/parameter_tests.dart';
+export 'src/test/redirect_tests.dart';
+export 'src/test/status_code_tests.dart';
