+++ dio_test/lib/util.dart
@@ -1,0 +1,2 @@
+export 'src/matcher.dart';
+export 'src/utils.dart';
