--- dio_tests/test/adapter/web.dart
@@ -1,6 +1,0 @@
-import 'package:dio/browser.dart';
-import 'package:dio/dio.dart';
-
-final List<HttpClientAdapter> adapters = [
-  BrowserHttpClientAdapter(),
-];
